started working on team creation and joining
The team db will be in memory for now.
A user can join a team and will be remembered by a cookie.
In the future the cookie could contain a player id. So that the backend can look for the users name per team.
For now the user will have only one name for all teams.